+++ src/common/documentationUrls.ts
@@ -1,0 +1,5 @@
+export const DOCS_URL_BASE = 'https://devops-extensions.dev';
+export const DOCS_URL_EXTENSION = `${DOCS_URL_BASE}/docs/extensions/auto-state`;
+export const DOCS_URL_FILTERS = `${DOCS_URL_EXTENSION}/rules/filters`;
+export const DOCS_URL_CHILD_LOOKUP = `${DOCS_URL_EXTENSION}/rules/children-lookup`;
+export const DOCS_URL_PROCESS_PARENT = `${DOCS_URL_EXTENSION}/rules/process-parent`;